requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.34
pydantic==2.9.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
telethon==1.35.0
apscheduler==3.10.4
httpx==0.27.0
aiosqlite==0.20.0
pytest==8.2.2

